developed an authentication function
Users have to login in with an email and a password before they can manipulate the info. The password is stored in database as a hashed style after creation. When login to the system, the backend will compare the hashed password, and it fails if they are different, but produces a token if success . And when the user does any query, the system will validate the token to make sure the user who is operating is still valid.